+++ Main/src/main/java/com/seveneleven/project/controller/ProjectChecklistDocs.java
@@ -21,6 +21,33 @@ import java.util.List;
 @RequestMapping("/api/projects")
 public interface ProjectChecklistDocs {
 
+    @GetMapping("/steps/{stepId}")
+    @Operation(
+            summary = "체크리스트 조회",
+            description = "특정 단계의 체크리스트를 조회합니다.",
+            responses = {
+                    @ApiResponse(
+                            responseCode = "200",
+                            description = "성공적으로 체크리스트를 반환했습니다.",
+                            content = @Content(
+                                    mediaType = "application/json",
+                                    schema = @Schema(implementation = GetStepChecklist.Response.class)
+                            )
+                    )
+            },
+            parameters = {
+                    @Parameter(
+                            name = "stepId",
+                            description = "체크리스트를 조회할 단계의 ID",
+                            required = true,
+                            example = "1"
+                    )
+            }
+    )
+    ResponseEntity<APIResponse<GetStepChecklist.Response>> getProjectChecklist(
+            @PathVariable Long stepId
+    );
+
     @PostMapping("")
     @Operation(
             summary = "체크리스트 생성",
